+++ src/scalar/builder/tests.rs
@@ -1,3 +1,8 @@
+// SPDX-License-Identifier: MIT
+// Copyright 2026 Tom F. <https://github.com/tomtom215/>
+// My way of giving something small back to the open source community
+// and encouraging more Rust development!
+
 use super::*;
 use crate::types::TypeId;
 use libduckdb_sys::{duckdb_data_chunk, duckdb_function_info, duckdb_vector};